+++ src/main/java/com/example/cafe/order/repository/OrderRepository.java
@@ -2,8 +2,14 @@ package com.example.cafe.order.repository;
 
 import com.example.cafe.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import java.util.Optional;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
+
+    @Query("select o from Order o join fetch o.orderItems where o.id = :orderId")
+    Optional<Order> findByIdWithOrderItems(@Param("orderId") Long orderId);
 }